Update example_02.cpp
Remove a bunch of indirection that are not needed and likely cause unneeded cache misses. A vector stores its elements contiguously in memory, this allows accesses to be in the same section of memory and therefore quite fast, when you store a pointer and allocate it separately that data is no longer contiguous and is likely slower as a result.
I'd be interested in how this impacts performance, often fragmented memory causes some pessimization.